moved edit distance code to own file, restructured project
